requirements.txt
python-dotenv
requests
tqdm
transformers
peft
optimum
torch
flask
supabase
